add app lock list guider

+++ src/com/leo/appmaster/cleanmemory/ProcessCleaner.java
@@ -2,11 +2,14 @@ package com.leo.appmaster.cleanmemory;
 
 import java.util.List;
 
+import com.leo.appmaster.utils.AppUtil;
 import com.leo.appmaster.utils.ProcessUtils;
 
 import android.app.ActivityManager;
 import android.app.ActivityManager.RunningAppProcessInfo;
 import android.content.Context;
+import android.content.pm.ApplicationInfo;
+import android.util.Log;
 
 public class ProcessCleaner {
 
@@ -15,9 +18,12 @@ public class ProcessCleaner {
 	private static ProcessCleaner mInstance;
 
 	private long mLastCleanTime;
-	private long mLastMemUsed;
+
 	private long mMemTotal;
 	private long mLastCleanMem;
+
+	private long mLastMemUsed;
+	private long mCurUsedMem;
 
 	private static int CLEAN_INTERVAL = 30 * 1000;
 
@@ -42,14 +48,17 @@ public class ProcessCleaner {
 	public long getLastCleanMem() {
 		return mLastCleanMem;
 	}
-	
+
 	public long getUsedMem() {
 		long curTime = System.currentTimeMillis();
 		if ((curTime - mLastCleanTime) > CLEAN_INTERVAL) {
 			mLastMemUsed = ProcessUtils.getUsedMem(mContext);
 		}
-
 		return mLastMemUsed;
+	}
+
+	public long getCurUsedMem() {
+		return mCurUsedMem;
 	}
 
 	public boolean allowClean() {
@@ -62,16 +71,14 @@ public class ProcessCleaner {
 		this.mContext = ctx.getApplicationContext();
 	}
 
-	public long tryClean() {
+	public long tryClean(Context ctx) {
 		long resault = -1;
 		long curTime = System.currentTimeMillis();
 		if ((curTime - mLastCleanTime) > CLEAN_INTERVAL) {
 			mLastCleanTime = curTime;
-			cleanAllProcess();
-			cleanAllProcess();
-			long curUsedMem = ProcessUtils.getUsedMem(mContext);
-			resault = mLastCleanMem = Math.abs(curUsedMem - mLastMemUsed);
-			mLastMemUsed = curUsedMem;
+			cleanAllProcess(ctx);
+			resault = mLastCleanMem = Math.abs(mCurUsedMem - mLastMemUsed);
+			mLastMemUsed = mCurUsedMem;
 		}
 		return resault;
 	}
@@ -101,10 +108,13 @@ public class ProcessCleaner {
 	/**
 	 * clean all running process
 	 */
-	public void cleanAllProcess() {
+	public void cleanAllProcess(Context cxt) {
 		List<RunningAppProcessInfo> list = mAm.getRunningAppProcesses();
 		for (RunningAppProcessInfo runningAppProcessInfo : list) {
-			mAm.killBackgroundProcesses(runningAppProcessInfo.processName);
+			if (runningAppProcessInfo.importance > RunningAppProcessInfo.IMPORTANCE_CANT_SAVE_STATE) {
+				mAm.killBackgroundProcesses(runningAppProcessInfo.processName);
+			}
 		}
+		mCurUsedMem = ProcessUtils.getUsedMem(cxt);
 	}
 }
